% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{Center.cell.MatD}
\alias{Center.cell.MatD}
\alias{Center.cell.MatD,MatD-method}
\title{Method Center.cell.MatD Centers all the cells of a matrix of distributions}
\usage{
Center.cell.MatD(object)

\S4method{Center.cell.MatD}{MatD}(object)
}
\arguments{
\item{object}{a MatD object, a matrix of distributions.}
}
\value{
A \code{MatD} object, having each distribution with a zero mean.
}
\description{
The function transform a MatD object (i.e. a matrix of distributions),
such that each distribution is shifted and has a mean equal to zero
}
\examples{
CEN_BLOOD <- Center.cell.MatD(BLOOD)
get.MatD.stats(BLOOD, stat = "mean")
}
