% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{[}
\alias{[}
\alias{[,MatD,ANY,ANY,ANY-method}
\alias{[,MatD-method}
\title{extract from a MatD Method [}
\usage{
\S4method{[}{MatD,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{matD} object}

\item{i}{a set of integer values identifying the rows}

\item{j}{a set of integer values identifying the columns}

\item{...}{not useful}

\item{drop}{a logical value inherited from the basic method "[" but not used (default=TRUE)}
}
\value{
A \code{MatD} object
}
\description{
This method overrides the "[" operator for a  \code{matD} object.
}
\examples{
D <- XXX # the BLOOD dataset
SUB_D <- XXX[c(1, 2, 5), c(1, 2)]
}
