% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{get.bars}
\alias{get.bars}
\alias{get.bars,distributionD-method}
\title{Method \code{get.bars}: show the distribution with bins}
\usage{
get.bars(object)

\S4method{get.bars}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object.}
}
\value{
A matrix: the two columns contains the values of DD and the the probablity (or the relative frequency) of the bin.
}
\description{
This functon return a data.frame describing the mass frequency of a \code{distributionD} object.
}
\examples{
D <- distributionD(x = c(1, 2, 3, 4), p = c(0.1, 0.2, 0.6, 1))
get.bars(D) # returns the discrete DD representation of D by a data.frame
}
