% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.cell.MatD}
\alias{get.cell.MatD}
\alias{get.cell.MatD,MatD,numeric,numeric-method}
\alias{get.cell.MatD,MatD-method}
\title{Method get.cell.MatD Returns the histogram in a cell of a matrix of distributions}
\usage{
get.cell.MatD(object, r, c)

\S4method{get.cell.MatD}{MatD,numeric,numeric}(object, r, c)
}
\arguments{
\item{object}{a MatD object, a matrix of distributions.}

\item{r}{an integer, the row index.}

\item{c}{an integer, the column index}
}
\value{
A \code{distributionD} object.
}
\description{
Returns the histogram data in the r-th row and the c-th column.
}
\examples{
get.cell.MatD(BLOOD, r = 1, c = 1)
}
