% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{get.m}
\alias{get.m}
\alias{get.m,distributionD-method}
\title{Method \code{get.m}: the mean of a distribution}
\usage{
get.m(object)

\S4method{get.m}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\value{
A numeric value
}
\description{
This functon return the mean of a \code{distributionD} object.
}
\examples{
D <- distributionD(x = c(1, 2, 3, 4), p = c(0.1, 0.2, 0.6, 1))
get.m(D) # returns the mean of D
}
