% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.main.info}
\alias{get.MatD.main.info}
\alias{get.MatD.main.info,MatD-method}
\title{Method get.MatD.main.info}
\usage{
get.MatD.main.info(object)

\S4method{get.MatD.main.info}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\value{
A list of char, the labels of the columns, or the names of the variables.
}
\description{
It returns the number of rows, of columns the labels of rows and columns of a \code{MatD} object.
}
\section{Slots}{

\describe{
\item{\code{nrows}}{\itemize{
\item the number of rows
}}

\item{\code{ncols}}{\itemize{
\item the number of columns
}}

\item{\code{rownames}}{\itemize{
\item a vector of char, the names of rows
}}

\item{\code{varnames}}{\itemize{
\item a vector of char, the names of columns
}}
}}

