% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.ncols}
\alias{get.MatD.ncols}
\alias{get.MatD.ncols,MatD-method}
\title{Method get.MatD.ncols}
\usage{
get.MatD.ncols(object)

\S4method{get.MatD.ncols}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\value{
An integer, the number of columns.
}
\description{
It returns the number of columns of a \code{MatD} object
}
