% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.nrows}
\alias{get.MatD.nrows}
\alias{get.MatD.nrows,MatD-method}
\title{Method get.MatD.nrows}
\usage{
\S4method{get.MatD.nrows}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\value{
An integer, the number of rows.
}
\description{
It returns the number of rows of a \code{MatD} object
}
