% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.rownames}
\alias{get.MatD.rownames}
\alias{get.MatD.rownames,MatD-method}
\title{Method get.MatD.rownames}
\usage{
get.MatD.rownames(object)

\S4method{get.MatD.rownames}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\value{
A vector of char, the label of the rows.
}
\description{
It returns the labels of the rows of a \code{MatD} object
}
