% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.stats}
\alias{get.MatD.stats}
\alias{get.MatD.stats,MatD-method}
\title{Method get.MatD.stats}
\usage{
get.MatD.stats(object, ...)

\S4method{get.MatD.stats}{MatD}(object, stat = "mean", prob = 0.5)
}
\arguments{
\item{object}{a \code{MatD} object}

\item{...}{a set of other parameters}

\item{stat}{(optional) a string containing the required statistic. Default='mean'\cr
\itemize{
\item \code{stat='mean'} - for computing the mean of each histogram\cr
\item \code{stat='median'} - for computing the median of each histogram\cr
\item \code{stat='min'} - for computing the minimum of each histogram\cr
\item \code{stat='max'} - for computing the maximum of each histogram\cr
\item \code{stat='std'} - for computing the standard deviatio of each histogram\cr
\item \code{stat='skewness'} - for computing the skewness of each histogram\cr
\item \code{stat='kurtosis'} - for computing the kurtosis of each histogram\cr
\item \code{stat='quantile'} - for computing the quantile ot level \code{prob} of each histogram\cr
}}

\item{prob}{(optional)a number between 0 and 1 for computing the value once choosen the \code{'quantile'} option for \code{stat}.}
}
\value{
A list
}
\description{
It returns statistics for each distribution contained in a \code{MatD} object.
}
\section{Slots}{

\describe{
\item{\code{stat}}{\itemize{
\item the chosen statistic
}}

\item{\code{prob}}{\itemize{
\item level of probability if stat='quantile'
}}

\item{\code{MAT}}{\itemize{
\item a matrix of values
}}
}}

\examples{
get.MatD.stats(BLOOD) # the means of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "median") # the medians of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "quantile", prob = 0.5) # the same as median
get.MatD.stats(BLOOD, stat = "min") # minima of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "quantile", prob = 0) # the same as min
get.MatD.stats(BLOOD, stat = "max") # maxima of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "quantile", prob = 1) # the same as max
get.MatD.stats(BLOOD, stat = "std") # standard deviations of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "skewness") # skewness indices of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "kurtosis") # kurtosis indices of the distributions in BLOOD dataset
get.MatD.stats(BLOOD, stat = "quantile", prob = 0.05)
# the fifth percentiles of distributions in BLOOD dataset
}
