% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{get.MatD.varnames}
\alias{get.MatD.varnames}
\alias{get.MatD.varnames,MatD-method}
\title{Method get.MatD.varnames}
\usage{
get.MatD.varnames(object)

\S4method{get.MatD.varnames}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\value{
A vector of char, the labels of the columns, or the names of the variables.
}
\description{
It returns the labels of the columns, or the names of the variables, of a \code{MatD} object
}
