% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{get.s}
\alias{get.s}
\alias{get.s,distributionD-method}
\title{Method \code{get.s}: the standard deviation of a distribution}
\usage{
get.s(object)

\S4method{get.s}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object.}
}
\value{
A numeric positive value, the standard deviation.
}
\description{
This functon return the standard deviation of a \code{distributionD} object.
}
\examples{
D <- distributionD(x = c(1, 2, 3, 4), p = c(0.1, 0.2, 0.6, 1))
get.s(D) # returns the standard deviation of D
}
