% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{kurtD}
\alias{kurtD}
\alias{kurtD,distributionD-method}
\title{Method \code{kurtD}: computes the kurthosis of a distribution}
\usage{
kurtD(object)

\S4method{kurtD}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\value{
A value for the kurtosis index, 3 is the kurtosis of a Gaussian
distribution
}
\description{
Kurtosis of a discrete DD (using the fourth standardized moment)
}
\examples{

## ---- A mydist distribution ----
mydist <- distributionD(x = c(1, 2, 3, 10), p = c(0.05, 0.1, 0.5, 1))
## ---- Compute the kurtosis of mydist ----
kurtD(mydist)
}
\author{
Antonio Irpino
}
\keyword{distribution}
