% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MatD}
\alias{MatD}
\alias{initialize,MatD-method}
\title{Constructor method for MatD class}
\usage{
\S4method{initialize}{MatD}(
  .Object,
  nrows = 1,
  ncols = 1,
  ListOfDist = NULL,
  names.rows = NULL,
  names.cols = NULL,
  by.row = FALSE
)
}
\description{
Constructor method for MatD class
}
