% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MatH-class}
\alias{MatH-class}
\title{Class MatH.}
\arguments{
\item{.Object}{the object type "MatD"}

\item{ListOfDist}{a vector or a list of \code{distributionD} objects}

\item{names.rows}{a vector or list of strings with thenames of the rows}

\item{names.cols}{a vector or list of strings with thenames of the columns (variables)}
}
\description{
Class \code{MatH} defines a matrix of \code{distributionH} objects
}
\examples{

## ---- create a list of six distributionD objects
ListOfDist <- vector("list", 6)
ListOfDist[[1]] <- distributionD(c(1, 2, 3), c(0.2, 0.4, 1))
ListOfDist[[2]] <- distributionD(c(7, 8, 10, 15), c(0.1, 0.2, 0.7, 1))
ListOfDist[[3]] <- distributionD(c(9, 11, 20), c(0.3, 0.5, 1))
ListOfDist[[4]] <- distributionD(c(2, 5, 8), c(0.1, 0.3, 1))
ListOfDist[[5]] <- distributionD(c(8, 10, 15), c(0.4, 0.75, 1))
ListOfDist[[6]] <- distributionD(c(20, 22, 24), c(0.05, 0.12, 1))

## create a MatH object filling it by columns
MyMAT <- new("MatD",
  nrows = 3, ncols = 2, ListOfDist = ListOfDist,
  names.rows = c("I1", "I2", "I3"), names.cols = c("Var1", "Var2"), by.row = FALSE
)

showClass("MatD")
}
\references{
Irpino, A., Verde, R. (2015) \emph{Basic
statistics for distributional symbolic variables: a new metric-based
approach} Advances in Data Analysis and Classification, DOI
10.1007/s11634-014-0176-4
}
\author{
Antonio Irpino
}
\keyword{classes}
