% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{meanD}
\alias{meanD}
\alias{meanD,distributionD-method}
\title{Method \code{meanD}: computes the mean of a distribution}
\usage{
meanD(object)

\S4method{meanD}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\value{
the mean of the distribution
}
\description{
Mean of a discrete  DD (First moment of the distribution)
}
\examples{

## ---- A mydist distribution ----
mydist <- distributionD(x = c(1, 2, 3, 10), p = c(0.05, 0.1, 0.5, 1))
## ---- Compute the mean of mydist ----
meanD(mydist)
}
\author{
Antonio Irpino
}
\keyword{distribution}
