% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\docType{methods}
\name{plot-distributionD}
\alias{plot-distributionD}
\alias{plot,distributionD-method}
\title{plot for a distributionD object}
\usage{
\S4method{plot}{distributionD}(x, type = "BAR", col = "green", border = "black", title = NULL)
}
\arguments{
\item{x}{a \code{distributionD} object}

\item{type}{(optional) a string describing the type of plot, default="BAR".\cr Other allowed types are
\cr"STAOR" stached horizontal \cr"STAVER" stacked vertical
\cr"CDF"=Cumulative distribution function, \cr"QF"= quantile function,
\cr"HBOXPLOT"=horizontal boxplot, \cr"VBOXPLOT"= vertical boxplot,}

\item{col}{(optional) a string the color of the plot, default="green".}

\item{border}{(optional) a string the color of the border of the plot, default="black".}
}
\description{
A plot function for a \code{distributionD} object. The function returns a representation
of the discrete DD.
}
\examples{
## ---- initialize a distributionD
mydist <- distributionD(x = c(7, 8, 10, 15), p = c(0.1, 0.2, 0.7, 1))
# show the discrete DD
plot(mydist) # plots mydist
plot(mydist, type = "BAR", col = "red", border = "blue") # plots mydist
plot(mydist, type = "STAOR", col = "red", border = "blue") # plots mydist stacked orizz
plot(mydist, type = "STAVER", col = "red", border = "blue") # plots mydist stacked vert
plot(mydist, type = "HBOXPLOT") # plots a horizontal boxplot for mydist
plot(mydist, type = "VBOXPLOT") # plots a vertical boxplot for mydist
plot(mydist, type = "CDF") # plots the cumulative distribution function of mydist
plot(mydist, type = "QF") # plots the quantile function of mydist
}
