% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\docType{methods}
\name{+}
\alias{+}
\alias{+,distributionD,distributionD-method}
\alias{+,numeric,distributionD-method}
\alias{+,distributionD,numeric-method}
\title{Method +}
\usage{
\S4method{+}{distributionD,distributionD}(e1, e2)

\S4method{+}{numeric,distributionD}(e1, e2)

\S4method{+}{distributionD,numeric}(e1, e2)
}
\arguments{
\item{e1}{a \code{distributionD} object or a number}

\item{e2}{a \code{distributionD} object or a number}
}
\value{
a \code{distributionD} object
}
\description{
the sum of two distribution according to the L2 Wassserstein

the sum of a number and a distribution according to the L2 Wasserstein

the sum of adistribution and a number according to the L2 Wassserstein
}
