% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{register}
\alias{register}
\alias{register,distributionD,distributionD-method}
\alias{register,distributionD-method}
\title{Method \code{register}}
\usage{
register(object1, object2)

\S4method{register}{distributionD,distributionD}(object1, object2)
}
\arguments{
\item{object1}{A \code{distributionD} object}

\item{object2}{A \code{distributionD} object}
}
\value{
The two \code{distributionD} objects in input sharing the same cdf (the \code{p}
slot)
}
\description{
Given two \code{distributionD} objects, it returns two equivalent distributions such that
they share the same cdf values. This function is useful for computing basic statistics.
}
\examples{

## ---- initialize two distributionD objects mydist1 and mydist2
mydist1 <- distributionD(c(1, 2, 3), c(0.1, 0.4, 1))
mydist2 <- distributionD(c(7, 8, 10, 15), c(0.1, 0.2, 0.7, 1))
## register the two distributions
regDist <- register(mydist1, mydist2)

## OUTPUT:
## regDist$[[1]]
## An object of class "distributionD"
## Slot "x": [1] 1.0 1.5 2.0 2.5 3.0
## Slot "p": [1] 0.1 0.2 0.4 0.7 1.0
## ...
## regDist$[[2]]
## An object of class "distributionD"
## Slot "x": [1] 7.0 8.0 8.8 10.0 15.0
## Slot "p": [1] 0.1 0.2 0.4  0.7  1.0
## ...
# The REGISTER function ----
}
\references{
Irpino, A., Lechevallier, Y. and Verde, R. (2006): \emph{Dynamic
clustering of histograms using Wasserstein metric} In: Rizzi, A., Vichi, M.
(eds.) COMPSTAT 2006. Physica-Verlag, Berlin, 869-876.\cr Irpino, A.,Verde,
R. (2006): \emph{A new Wasserstein based distance for the hierarchical
clustering of histogram symbolic data} In: Batanjeli, V., Bock, H.H.,
Ferligoj, A., Ziberna, A. (eds.) Data Science and Classification, IFCS 2006.
Springer, Berlin, 185-192.
}
\author{
Antonio Irpino
}
\keyword{distribution}
