% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{rQQ}
\alias{rQQ}
\alias{rQQ,distributionD,distributionD-method}
\alias{rQQ,distributionD-method}
\title{Method \code{rQQ}}
\usage{
rQQ(e1, e2)

\S4method{rQQ}{distributionD,distributionD}(e1, e2)
}
\arguments{
\item{e1}{A \code{distributionD} object}

\item{e2}{A \code{distributionD} object}
}
\value{
Pearson correlation index between quantiles
}
\description{
Quantile-Quantile correlation between two distributions
}
\examples{

## ---- initialize two distributionD object mydist1 and mydist2
mydist1 <- distributionD(x = c(1, 2, 3), p = c(0.1, 0.4, 1))
mydist2 <- distributionD(x = c(7, 8, 10, 15), p = c(0.05, 0.2, 0.7, 1))
## computes the rQQ
rQQ(mydist1, mydist2)
}
\references{
Irpino, A., Verde, R. (2015) \emph{Basic
statistics for distributional symbolic variables: a new metric-based
approach} Advances in Data Analysis and Classification, DOI 10.1007/s11634-014-0176-4
}
\author{
Antonio Irpino
}
