% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{set.cell.MatD}
\alias{set.cell.MatD}
\alias{set.cell.MatD,distributionD,MatD,numeric,numeric-method}
\alias{set.cell.MatD,MatD-method}
\title{Method set.cell.MatD assign a histogram to a cell of a matrix of histograms}
\usage{
set.cell.MatD(object, mat, r, c)

\S4method{set.cell.MatD}{distributionD,MatD,numeric,numeric}(object, mat, r, c)
}
\arguments{
\item{object}{a distributionD object, a matrix of distributions.}

\item{mat}{a MatD object, a matrix of distributions.}

\item{r}{an integer, the row index.}

\item{c}{an integer, the column index}
}
\value{
A \code{MatD} object.
}
\description{
Assign a histogram data to the r-th row and the c-th column of a matrix of histograms.
}
\examples{
mydist <- distributionD(x = c(0, 1, 2, 3, 4), p = c(0, 0.1, 0.6, 0.9, 1))
MAT <- set.cell.MatD(mydist, BLOOD, r = 1, c = 1)
}
