% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,distributionD-method}
\title{Method show for distributionD}
\usage{
\S4method{show}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\description{
An overriding show function for a \code{distributionD} object. The function returns a representation
of the discrete DD, if the number of bins is high the central part of the discrete DD is truncated.
}
\examples{
## ---- initialize a distributionD
mydist <- distributionD(x = c(7, 8, 10, 15), p = c(0, 0.2, 0.7, 1))
# show the discrete DD
mydist
}
