% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\docType{methods}
\name{show-MatD}
\alias{show-MatD}
\alias{show,MatD-method}
\title{Method show for MatD}
\usage{
\S4method{show}{MatD}(object)
}
\arguments{
\item{object}{a \code{MatD} object}
}
\description{
An overriding show method for a \code{MatD} object. The method returns a representation
of the matrix using the mean and the standard deviation for each histogram.
}
\examples{
show(BLOOD)
print(BLOOD)
BLOOD
}
