% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{skewD}
\alias{skewD}
\alias{skewD,distributionD-method}
\title{Method \code{skewD}: computes the skewness of a distribution}
\usage{
skewD(object)

\S4method{skewD}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\value{
A value for the skewness index
}
\description{
Skewness of a discrete DD (using the third standardized moment)
}
\examples{

## ---- A mydist distribution ----
mydist <- distributionD(x = c(1, 2, 3, 10), p = c(0.05, 0.1, 0.5, 1))
## ---- Compute the skewness of mydist ----
skewD(mydist)
}
\author{
Antonio Irpino
}
\keyword{distribution}
