% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{stdD}
\alias{stdD}
\alias{stdD,distributionD-method}
\title{Method \code{stdD}: computes the standard deviation of a distribution}
\usage{
stdD(object)

\S4method{stdD}{distributionD}(object)
}
\arguments{
\item{object}{a \code{distributionD} object}
}
\value{
A value for the standard deviation
}
\description{
Standard deviation of a discrete DD (i.e., the square root of the centered
second moment)
}
\examples{

## ---- A mydist distribution ----
mydist <- distributionD(x = c(1, 2, 3, 10), p = c(0.05, 0.1, 0.5, 1))
## ---- Compute the standard deviation of mydist ----
stdD(mydist)
}
\author{
Antonio Irpino
}
\keyword{distribution}
