% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{*-methods}
\alias{*-methods}
\alias{*,distributionD,distributionD-method}
\alias{*,numeric,distributionD-method}
\alias{*,distributionD,numeric-method}
\title{Method *}
\usage{
\S4method{*}{distributionD,distributionD}(e1, e2)

\S4method{*}{numeric,distributionD}(e1, e2)

\S4method{*}{distributionD,numeric}(e1, e2)
}
\arguments{
\item{e1}{a \code{distributionD} object or a number}

\item{e2}{a \code{distributionD} object or a number}
}
\description{
the product of a number and a distribution according to the L2 Wasserstein

the product of a number and a distribution according to the L2 Wasserstein

the product of a number and a distribution according to the L2 Wasserstein
}
