% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_distD.R
\name{WassSqDistD}
\alias{WassSqDistD}
\alias{WassSqDistD,distributionD,distributionD-method}
\alias{WassSqDistD,distributionD-method}
\title{Method \code{WassSqDistD}}
\usage{
WassSqDistD(object1, object2, ...)

\S4method{WassSqDistD}{distributionD,distributionD}(object1 = object1, object2 = object2, details = FALSE)
}
\arguments{
\item{object1}{is an object of \env{distributionD} class}

\item{object2}{is an object of \env{distributionD} class}

\item{...}{optional parameters}

\item{details}{(optional, default=FALSE) is a logical value, if TRUE returns the decomposition of the distance}
}
\value{
If \code{details=FALSE}, the function returns the squared L2 Wasserstein distance.\cr
If \code{details=TRUE}, the function returns list containing the squared distance, its
decomposition in three parts (position, size and shape) and the correlation coefficient between the quantile functions.
}
\description{
Computes the squared L2 Wasserstein distance between two \code{distributionD} objects.
}
\examples{
## ---- create two distributionD objects ----
mydist1 <- distributionD(x = c(1, 2, 3), p = c(0.1, 0.4, 1))
mydist2 <- distributionD(x = c(7, 8, 10, 15), p = c(0.1, 0.2, 0.7, 1))
# -- compute the squared L2 Waaserstein distance
WassSqDistD(mydist1, mydist2)
# -- compute the squared L2 Waaserstein distance with details
WassSqDistD(mydist1, mydist2, details = TRUE)
}
\references{
Irpino, A. and Romano, E. (2007): \emph{Optimal histogram representation of large data sets:
Fisher vs piecewise linear approximations}. RNTI E-9, 99-110.\cr
Irpino, A., Verde, R. (2015) \emph{Basic
statistics for distributional symbolic variables: a new metric-based
approach} Advances in Data Analysis and Classification, DOI 10.1007/s11634-014-0176-4
}
\keyword{distribution}
