% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.bind}
\alias{WH.bind}
\alias{WH.bind,MatD,MatD-method}
\alias{WH.bind,MatD-method}
\title{Method WH.bind}
\usage{
WH.bind(object1, object2, byrow)

\S4method{WH.bind}{MatD,MatD}(object1, object2, byrow = TRUE)
}
\arguments{
\item{object1}{a \code{MatD} object}

\item{object2}{a \code{MatD} object}

\item{byrow}{a logical value (default=TRUE) attaches the objects by row}
}
\value{
a \code{MatD} object,
}
\description{
It attaches two \code{MatD} objects with the same columns by row, or the same rows by colum.
}
\examples{
# binding by row
M1 <- BLOOD[1:10, 1]
M2 <- BLOOD[1:10, 3]
MAT <- WH.bind(M1, M2, byrow = TRUE)
# binding by col
M1 <- BLOOD[1:10, 1]
M2 <- BLOOD[1:10, 3]
MAT <- WH.bind(M1, M2, byrow = FALSE)
}
\seealso{
\code{\link{WH.bind.row}} for binding by row, \code{\link{WH.bind.col}} for binding by column
}
