% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.bind.col}
\alias{WH.bind.col}
\alias{WH.bind.col,MatD,MatD-method}
\alias{WH.bind.col,MatD-method}
\title{Method WH.bind.col}
\usage{
WH.bind.col(object1, object2)

\S4method{WH.bind.col}{MatD,MatD}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatD} object}

\item{object2}{a \code{MatD} object}
}
\value{
a \code{MatD} object,
}
\description{
It attaches two \code{MatD} objects with the same rows by colums.
}
\examples{
M1 <- BLOOD[1:10, 1]
M2 <- BLOOD[1:10, 3]
MAT <- WH.bind.col(M1, M2)
}
