% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.bind.row}
\alias{WH.bind.row}
\alias{WH.bind.row,MatD,MatD-method}
\alias{WH.bind.row,MatD-method}
\title{Method WH.bind.row}
\usage{
WH.bind.row(object1, object2)

\S4method{WH.bind.row}{MatD,MatD}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatD} object}

\item{object2}{a \code{MatD} object}
}
\value{
a \code{MatD} object,
}
\description{
It attaches two \code{MatD} objects with the same columns by row.
}
\examples{
M1 <- BLOOD[1:3, ]
M2 <- BLOOD[5:8, ]
MAT <- WH.bind.row(M1, M2)
}
