% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.mat.prod}
\alias{WH.mat.prod}
\alias{WH.mat.prod,MatD,MatD-method}
\alias{WH.mat.prod,MatD-method}
\title{Method WH.mat.prod}
\usage{
WH.mat.prod(object1, object2, ...)

\S4method{WH.mat.prod}{MatD,MatD}(object1, object2, traspose1 = FALSE, traspose2 = FALSE)
}
\arguments{
\item{object1}{a \code{MatD} object}

\item{object2}{a \code{MatD} object}

\item{...}{other optional parameters}

\item{traspose1}{a logical value, default=FALSE. If TRUE trasposes object1}

\item{traspose2}{a logical value, default=FALSE. If TRUE trasposes object2}
}
\value{
a matrix of numbers
}
\description{
It is the matrix product of two \code{MatD} objects, i.e. two matrices of distributions,
by using the dot product of two histograms that is consistent with
a set of distributions equipped with a L2 wasserstein metric.
}
\examples{

M1 <- BLOOD[1:5, ]
M2 <- BLOOD[6:10, ]
MAT <- WH.mat.prod(M1, M2, traspose1 = TRUE, traspose2 = FALSE)
}
