% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.mat.sum}
\alias{WH.mat.sum}
\alias{WH.mat.sum,MatD,MatD-method}
\alias{WH.mat.sum,MatD-method}
\title{Method WH.mat.sum}
\usage{
WH.mat.sum(object1, object2)

\S4method{WH.mat.sum}{MatD,MatD}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatD} object}

\item{object2}{a \code{MatD} object}
}
\value{
a \code{MatD} object,
}
\description{
It sums two \code{MatD} objects, i.e. two matrices of distributions,
by summing the quantile functions of histograms. This sum is consistent with
a set of distributions equipped with a L2 wasserstein metric.
}
\examples{
# binding by row
M1 <- BLOOD[1:5, ]
M2 <- BLOOD[6:10, ]
MAT <- WH.mat.sum(M1, M2)
}
