% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.vec.mean}
\alias{WH.vec.mean}
\alias{WH.vec.mean,MatD-method}
\title{Method WH.vec.mean}
\usage{
WH.vec.mean(object, ...)

\S4method{WH.vec.mean}{MatD}(object, w = numeric(0))
}
\arguments{
\item{object}{a \code{MatD} object}

\item{...}{optional arguments}

\item{w}{it is possible to add a vector of weights (positive numbers) having the same size of
the \code{MatD object}, default = equal weights for all}
}
\value{
a \code{distributionD} object, i.e. a histogram
}
\description{
Compute a histogram that is the weighted mean of the set of histograms contained
in a \code{MatD} object, i.e. a matrix of histograms, consistent with
a set of distributions equipped with a L2 wasserstein metric.
}
\examples{
hmean <- WH.vec.mean(BLOOD)
# generate a set of random weights
RN <- runif(get.MatD.nrows(BLOOD) * get.MatD.ncols(BLOOD))
hmean <- WH.vec.mean(BLOOD, w = RN)
}
