% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatD.R
\name{WH.vec.sum}
\alias{WH.vec.sum}
\alias{WH.vec.sum,MatD-method}
\title{Method WH.vec.sum}
\usage{
WH.vec.sum(object, ...)

\S4method{WH.vec.sum}{MatD}(object, w = numeric(0))
}
\arguments{
\item{object}{a \code{MatD} object}

\item{...}{optional arguments}

\item{w}{it is possible to add a vector of weights (positive numbers) having the same size of the \code{MatD object},
default = equal weights for all cells}
}
\value{
a \code{distributionD} object, i.e. a histogram
}
\description{
Compute a histogram that is the weighted sum of the set of histograms contained
in a \code{MatD} object, i.e. a matrix of histograms, consistent with
a set of distributions equipped with a L2 wasserstein metric.
}
\examples{
hsum <- WH.vec.sum(BLOOD)
# generate a set of random weights
RN <- runif(get.MatD.nrows(BLOOD) * get.MatD.ncols(BLOOD))
hsum <- WH.vec.sum(BLOOD, w = RN)
### SUM of distributions ----
}
